Adds support for parallel testing
